Moved hyperopt functions around
